scraper/requirements.txt
requests
beautifulsoup4
pandas
pymongo
python-dotenv
pythainlp
tenacity
